Updating calc to account for taa stride changes.
Biggest change is filtering the abd and chest signals before passing
them to the taa function. This is easier since the output of the taa is
no longer in seconds but depends on the breathing frequency of the
signal.

Read and calculate over the baseline and followup and include both in
the results.